rename common to cc_common
to reduce the risk of namespace collisions
