+++ backend/compact-connect/lambdas/common-python/cc_common/config.py
@@ -26,7 +26,7 @@ class _Config:
 
     @cached_property
     def data_client(self):
-        from common.data_model.client import DataClient
+        from cc_common.data_model.client import DataClient
 
         return DataClient(self)
