+++ backend/compact-connect/lambdas/common-python/cc_common/data_model/client.py
@@ -6,11 +6,11 @@ from boto3.dynamodb.conditions import Attr, Key
 from boto3.dynamodb.types import TypeDeserializer, TypeSerializer
 from botocore.exceptions import ClientError
 
-from common.config import _Config, config, logger
-from common.data_model.query_paginator import paginated_query
-from common.data_model.schema import PrivilegeRecordSchema
-from common.data_model.schema.base_record import SSNIndexRecordSchema
-from common.exceptions import CCNotFoundException
+from cc_common.config import _Config, config, logger
+from cc_common.data_model.query_paginator import paginated_query
+from cc_common.data_model.schema import PrivilegeRecordSchema
+from cc_common.data_model.schema.base_record import SSNIndexRecordSchema
+from cc_common.exceptions import CCNotFoundException
 
 
 class DataClient:
