+++ backend/compact-connect/lambdas/common-python/cc_common/data_model/query_paginator.py
@@ -6,9 +6,9 @@ from types import MethodType
 from botocore.exceptions import ClientError
 from marshmallow import ValidationError
 
-from common.config import config, logger
-from common.data_model.schema.base_record import BaseRecordSchema
-from common.exceptions import CCInternalException, CCInvalidRequestException
+from cc_common.config import config, logger
+from cc_common.data_model.schema.base_record import BaseRecordSchema
+from cc_common.exceptions import CCInternalException, CCInvalidRequestException
 
 
 # It's conventional to name a decorator in snake_case, even if it is implemented as a class
