+++ backend/compact-connect/lambdas/common-python/cc_common/data_model/schema/__init__.py
