+++ backend/compact-connect/lambdas/common-python/cc_common/data_model/schema/base_record.py
@@ -8,7 +8,7 @@ from marshmallow import EXCLUDE, RAISE, Schema, post_load, pre_dump
 from marshmallow.fields import UUID, Date, List, String
 from marshmallow.validate import Regexp
 
-from common.exceptions import CCInternalException
+from cc_common.exceptions import CCInternalException
 
 
 class StrictSchema(Schema):
