+++ backend/compact-connect/lambdas/common-python/cc_common/exceptions.py
