+++ backend/compact-connect/lambdas/provider-data-v1/tests/__init__.py
@@ -28,8 +28,8 @@ class TstLambdas(TestCase):
         )
         # Monkey-patch config object to be sure we have it based
         # on the env vars we set above
-        import common.config
+        import cc_common.config
 
-        cls.config = common.config._Config()  # noqa: SLF001 protected-access
-        common.config.config = cls.config
+        cls.config = cc_common.config._Config()  # noqa: SLF001 protected-access
+        cc_common.config.config = cls.config
         cls.mock_context = MagicMock(name='MockLambdaContext', spec=LambdaContext)
